refactor: move format methods to core
