+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
+export * from './format';
 export * from './key_pair';
 export * from './key_store';
 export * from './provider';
